refactor(tier4_planning_rviz_plugin): clean up the code of path
